ignore find exit status, fix recursive removing
The find command exits with error when it encounters a problem e.g. with access rights. It won't write those to stdout so we can ignore these and process valid output.

Secondly, in a situation like the following, the tool failed after removing the containing folder, when it tried to remove the subdirectories:

  "./grade-school/.stack-work",
  "./grade-school/.stack-work/dist/x86_64-osx/Cabal-2.2.0.1/build/test/test-tmp/.stack-work",
  "./grade-school/.stack-work/dist/x86_64-osx/Cabal-2.2.0.1/build/.stack-work",